ShouldSetupComponentStyles -> PASS
