+++ DMX.Portal.Web/Views/Components/NewLabDialogs/NewLabDialog.razor.Styles.cs
@@ -1,0 +1,32 @@
+﻿// --------------------------------------------------------------- 
+// Copyright (c) Microsoft Corporation. All rights reserved. 
+// ---------------------------------------------------------------
+
+using DMX.Portal.Web.Models.Views.Components.NewLabDialogComponents;
+using DMX.Portal.Web.Views.Bases;
+using SharpStyles.Models;
+
+namespace DMX.Portal.Web.Views.Components.NewLabDialogs
+{
+    public partial class NewLabDialog
+    {
+        public StyleBase StyleElement { get; set; }
+        public NewLabDialogComponentStyle Style { get; set; }
+
+        private void SetupStyle()
+        {
+            this.Style = new NewLabDialogComponentStyle
+            {
+                NewLabTextbox = new SharpStyle
+                {
+                    PaddingBottom = "20px"
+                },
+
+                NewLabErrorMessage = new SharpStyle
+                {
+                    Color = "red"
+                }
+            };
+        }
+    }
+}
