+++ DMX.Portal.Web/Views/Components/NewLabDialogs/NewLabDialog.razor.cs
@@ -9,7 +9,6 @@ using DMX.Portal.Web.Models.Views.LabViews;
 using DMX.Portal.Web.Services.Views.LabViews;
 using DMX.Portal.Web.Views.Bases;
 using Microsoft.AspNetCore.Components;
-using SharpStyles.Models;
 
 namespace DMX.Portal.Web.Views.Components.NewLabDialogs
 {
@@ -19,8 +18,6 @@ namespace DMX.Portal.Web.Views.Components.NewLabDialogs
         public ILabViewService LabViewService { get; set; }
 
         public NewLabDialogComponentState State { get; set; }
-        public StyleBase StyleElement { get; set; }
-        public NewLabDialogComponentStyle Style { get; set; }
         public DialogBase Dialog { get; set; }
         public TextBoxBase LabId { get; set; }
         public TextBoxBase LabDmxVersion { get; set; }
@@ -84,21 +81,6 @@ namespace DMX.Portal.Web.Views.Components.NewLabDialogs
         {
             this.Dialog.Hide();
             this.IsVisible = this.Dialog.IsVisible;
-        }
-
-        private void SetupStyle()
-        {
-            this.Style = new NewLabDialogComponentStyle
-            {
-                NewLabTextBox = new SharpStyle
-                {
-                    PaddingBottom = "20px"
-                },
-                NewLabErrorMessage = new SharpStyle
-                {
-                    Color = "red"
-                }
-            };
         }
     }
 }
