--- DMX.Portal.Web/Views/Components/NewLabDialogs/NewLabDialog.razor.css
@@ -1,7 +1,0 @@
-﻿::deep .new-lab-textbox {
-    padding-bottom: 20px;
-}
-
-::deep .new-lab-error-message {
-    color: red;
-}
